Extract json_selection/helpers.rs from parser.rs.
