+++ apollo-federation/src/sources/connect/json_selection/mod.rs
@@ -1,2 +1,4 @@
+mod helpers;
 mod parser;
+
 pub use parser::*;
